+++ Runtime/Common.cs
@@ -1,0 +1,31 @@
+﻿using UnityEngine;
+
+internal struct Boid
+{
+    public Vector3 position; 
+    public Vector3 direction;
+    public const int size = 24;
+} 
+
+internal struct ObstacleData
+{
+    public Vector3 position; 
+    public float radius;
+    public const int size = 16;
+}
+
+internal struct TargetData
+{
+    public Vector3 position;
+    public float weight; // Encoded in the .w component of a float4 in shader
+    public const int size = 16;
+}
+
+internal struct ZoneData
+{
+    public Matrix4x4 worldToLocal;
+    public Matrix4x4 localToWorld;
+    public Vector3 dimensions;
+    public int type;
+    public const int size = 144; // 64+64+12+4 = 144 bytes
+}